pyproject.toml
[project]
name = "agriculture-iot-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.77",
    "openpyxl>=3.1.5",
    "pandas<3.0.0",
    "python-dotenv",
    "pyyaml",
    "pytest",
    "awswrangler>=3.0.0",
    "pydantic>=2.0.0",
    "python-json-logger>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pandas-stubs",
    "types-PyYAML",
]
